Add fractional FPS support.
Now that we put the timing info into the bitstream, the time base must
be precisely known. Represent framerate as a fraction and add timing
info only if the old floating point framerate was not used.

Deprecate cfg->framerate so it can be removed once we get patches to
FFmpeg and libav.

Add support for (num)/(denom) format to --input-fps.