Fix autoreconf and configure without pkg-config
The PKG_CHECK_MODULES macro is provided by pkg-config. Not having the
macro available caused autoreconf and configure to fail with obscure
error messages. This commit adds the file pkg.m4 from pkg-config so that
the macros are available even when pkg-config is not installed. Now
pkg-config is only needed when running configure --with-cryptopp.

Fixes #138.