Fix Gitlab CI tests
Drops the build stage in Gitlab CI tests and includes build in the test
jobs. Building in a separate stage did not work as intended. The test
jobs rebuilt Kvazaar without the CFLAGS given in the build jobs and
overwrote the results from the build jobs. The test were therefore run
without the sanitizers enabled.