Do not build AVX2 with GCC on MinGW
There is apparently no way to build non-inline functions accepting or
returning AVX2 params with GCC on MinGW without mangling the source code
for all platforms, because it simply refuses to align stack to more than
16 bytes where needed. Drop the AVX2 optimizations on that platform to
have a working build.

AVX2 for MinGW builds is still supported with Clang, please use it
instead because it works.